AI/mcp/jira-server/pyproject.toml
[project]
name = "jira-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.2",
]
